Repourpose and rewrite DoubleSidedLineTraceMultiByChannelWithPenetrations()
Now outputs array of new struct indicating if it was an entrnance/exit,
alleviating the need to figure out if it was entrance/exit after the
call.
Make exit hit result distances and time values relative to the forwards trace to allow for easy and consistant usage of outputed hit results.
Optimization is now optional through a bool parameter since the
optimization turns out to not work in specific scenarios.
Cleanup code