Account for sweep shapes for the backwards scene cast start.
We achieve this by using UBFL_MathHelpers::GetCollisionShapeBoundingSphereRadius() and giving that bounding radius to UBFL_CollisionQueryHelpers::DetermineBackwardsSceneCastStart() for it to use as padding for the optimization.

Also fix DetermineBackwardsSceneCastStart() by stopping letting it go past the InForwardsEnd. Before it was bumping forwards past it by the avoidance padding but that doesn't make sense to do actually. Now it doesn't do that.

Also clean up DetermineBackwardsSceneCastStart() code and comments.

Have PenetrationSweepWithExitHits() and SweepMultiWithExitHits() use the sweep shape bounding radius.