Make scene cast functions that are generic to linetraces and sweeps.
This is possible by using the engine's FCollisionShape with ECollisionShape::Line to indicate a line trace. Other ECollisionShape types use the sweep method.
This reduces the amount of duplicate code a lot. Functions that were
once used to repeat the same logic as other functions in order to
achieve slightly differnt functionality are now a single line, calling
on the new generic scene cast functions.